Adding some modifications to either dm header service and puppet-master

+++ site/profile/manifests/dm/dm_header_service.pp
@@ -10,9 +10,14 @@ class profile::dm::dm_header_service{
 		ts_sal_path => $ts_sal_path,
 	}
 
+	package { 'numpy':
+		ensure => installed,
+	}
+
 	exec { 'get-custom-fitsio':
 		command => 'wget https://github.com/menanteau/fitsio/archive/master.tar.gz -O /tmp/fitsio-master.tar.gz && cd /tmp/',
 		path => '/bin/',
+		onlyif => "test ! -f /tmp/fitsio-master.tar.gz"
 	}
 	
 	exec{ 'install-custom-fitsio':
